containers/migrate/requirements.txt
python-redmine
mysql-connector-python
